Часть интерфейса
